creates item controller
